paged-kv: final review hardening (int64 op signatures, dtype-safe mask, docs)
- Declare paged op block_idx/offset_idx as INT64 to match what the cache layer
  feeds (and the int64 head Range), removing the signature/feed inconsistency.
- v_out invalid-position masking uses v_out.dtype instead of hardcoded float32,
  so fp16/bf16/fp8 KV is not silently up/down-cast (paged path).
- Document the runtime contract on QEffPagedDynamicLayer: past_key_values are the
  block-pool tensors, the last block is a reserved null block (size pool +1), and
  block_table entries must stay in [0, num_blocks).

Full suite green; eager-vs-onnxruntime ops numeric diff = 0.0.

+++ QEfficient/customop/ctx_paged_scatter_gather.py
@@ -37,8 +37,8 @@ ops = getattr(onnxscript, "opset" + str(constants.ONNX_EXPORT_OPSET))
 @onnxscript.script(onnxscript.values.Opset("com.qualcomm.cloud", 1))
 def CtxScatterPaged(
     data: onnxscript.FLOAT,
-    block_idx: onnxscript.INT32,
-    offset_idx: onnxscript.INT32,
+    block_idx: onnxscript.INT64,
+    offset_idx: onnxscript.INT64,
     updates: onnxscript.FLOAT,
 ) -> onnxscript.FLOAT:
     # data: [num_blocks, num_heads, page_size, head_dim]
@@ -99,8 +99,8 @@ class CtxScatterPagedFunc(torch.autograd.Function):
 @onnxscript.script(onnxscript.values.Opset("com.qualcomm.cloud", 1))
 def CtxGatherPaged(
     data: onnxscript.FLOAT,
-    block_idx: onnxscript.INT32,
-    offset_idx: onnxscript.INT32,
+    block_idx: onnxscript.INT64,
+    offset_idx: onnxscript.INT64,
 ) -> onnxscript.FLOAT:
     # data: [num_blocks, num_heads, page_size, head_dim]
     # block_idx, offset_idx: [batch, ctx]   -> out: [batch, num_heads, ctx, head_dim]
